* main: (29 commits)
  ftp --> resources
  Update release notes
  Convert the ENABLE_XXX options to NETCDF_ENABLE_XXX options
  Modify CMakeLists.txt to honor CMAKE_INSTALL_MANDIR in support of https://github.com/Unidata/netcdf-c/issues/2920.
  Modify ncdump to print char-valued variables as utf8.
  CI: Create an MSYS2/MinGW CMake run.
  CI: Test on MSYS2/UCRT64 environment
  CMake: Fix running tests on MinGW
  removal of ftp and contrib site
  removal of ftp site
  removal of ftp and netcdf contrib site
  removal of ftp and netcdf contrib site
  comment
  ftp --> resources for sample data location
  ftp--> resources for sample data location
  CMake: Enable plugins on MinGW
  Reintroduce targets to allow for cmake-based netCDF-Fortran to find and linka gainst netCDF-C. Stop-gap measure until we can modify netCDF-Fortran to use a more modern approach.
  revert
  Fix conversion warning in flag set/check/clear macros
  ckp
  ...

+++ libdap4/ncd4types.h
@@ -13,6 +13,7 @@ are defined here.
 
 #undef COMPILEBYDEFAULT
 
+#include "ncdap.h"
 #include "ncrc.h"
 #include "ncauth.h"
 
@@ -206,10 +207,10 @@ struct NCD4node {
         NClist* values;
     } attr;
     struct { /* sort == NCD4_OPAQUE */
-	long long size; /* 0 => var length */
+        size_t size; /* 0 => var length */
     } opaque;
     struct { /* sort == NCD4_DIMENSION */
-	long long size;
+	size_t size;
 	int isunlimited;
 	int isanonymous;
     } dim;
